Modified README for 2015...
whiskeyDem was last compiled in 2005 with java and java3D.  It may be
difficult to compile the code with modern tools.  In a nutshell, this
program will create a 3d virtual world with an aircraft as an avatar.
There are no flight model physics, bounding box edge detection or video
game mechanics implemented.  Very basic networking abilities are
included
that was used to test various networking concepts to provide the remote
user with a "good" experience with minimal network lag, etc.  The
scenery
generated is provided by the USGS Digital Elevation Maps.  Feedback
welcome
especially if the project can compile with modern java tools!